membase_1.6.0beta2 formula
Conflicts with the existing memcached, membase, and libmemcached
forumla; we ensure these are removed before installing.

This forumla has been tested by the membase community here:

http://groups.google.com/group/membase/t/69a47f80a2b772b8

Future stable releases of membase will be better packaged as
components to allow for a simpler forumla  and to remove conflicts.
For now, this is the only way to get a stable build.